+++ frontend/src/app/events/page.tsx
@@ -1,5 +1,5 @@
 import { Card } from '../components/events/post-card'
-import { fetchEvents } from '../lib/data'
+import { fetchEvents } from '../../lib/data'
 
 export default async function Component() {
   const posts = await fetchEvents()
